chore(chat): split tool defs into per-domain modules (#344)
Pure refactor, zero behaviour change. The 23 `tool()` blocks that used
to live inline in web/src/app/api/chat/route.ts now live under
web/src/lib/tools/, one file per domain, exported as factory functions
that take a `ToolContext` and return the tool object.

Modules:
- tasks, habits, fitness, profile (data-access, no external deps)
- gmail (owns DEMO_EMAILS fixture)
- calendar (owns DEMO_CALENDAR_EVENTS fixture)
- meals (get_recipes, get_today_meals)
- session (get_session_history — split out of workouts because its
  dependency is the current chat session, not a data domain)
- workouts (5 tools; owns WorkoutExercise, buildCalendarDescription,
  withTimeout helpers that were previously private to POST)
- stocks, sports

Shared:
- _contract.ts — ok()/err()/OkResult<T>/ErrResult (the #319
  verified-success contract)
- _context.ts — ToolContext interface

Stays in route.ts:
- isToolResultOk(), synthesizeFallbackSummary(), CompletedToolStep —
  turn-orchestration concerns, not tool contract
- retryOnOverload middleware, selectModel, system prompts, the POST
  handler itself with its streamText/onFinish/messageMetadata wiring

route.ts: 1954 lines → 576 lines.

Typecheck + build green.
