+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ## [Unreleased]
 
 ### Changed
+- **Chat tool definitions split out of `route.ts` into per-domain modules (#344).** The 23 `tool()` blocks that used to live inline in [web/src/app/api/chat/route.ts](web/src/app/api/chat/route.ts) — tasks, habits, fitness, profile, gmail, calendar, meals, session, workouts, stocks, sports — each now live in their own module under [web/src/lib/tools/](web/src/lib/tools/), exported as factory functions (`buildTasksTools(ctx)` etc.) that take a `ToolContext` ({ supabase, userId, isDemo, sessionId }) and return the tool object. Pure refactor, zero behaviour change; `route.ts` shrunk from 1954 lines to 576 (tool definitions: 1404 lines → 0 inline). Shared: `web/src/lib/tools/_contract.ts` holds the `ok()` / `err()` helpers + `OkResult<T>` / `ErrResult` types (the #319 verified-success contract); `web/src/lib/tools/_context.ts` defines the per-request context. `isToolResultOk()` and `synthesizeFallbackSummary()` stay in `route.ts` because they're turn-orchestration concerns, not tool contract. Demo fixtures (`DEMO_EMAILS`, `DEMO_CALENDAR_EVENTS`) live in their respective tool modules; workout helpers (`WorkoutExercise`, `buildCalendarDescription`, `withTimeout`) live in [workouts.ts](web/src/lib/tools/workouts.ts).
 - **Vercel AI SDK upgraded v4 → v6 (#329).** Two-leg staged migration (4 → 5, validated; 5 → 6, validated). Dependency bumps: `ai@^4.3 → ^6.0`, `@ai-sdk/anthropic@^1.2 → ^3.0`, `@ai-sdk/groq@^1.2 → ^3.0`, added `@ai-sdk/react@^3.0` (v5 moved `useChat` out of `ai/react`), `@ai-sdk/provider@^3.0`, `zod@^4.1.8`. No user-facing behaviour change; the rewrite is internal plumbing so future upgrades don't accumulate debt on top of #319's trickiest code. Key migrations: (a) `tool({ parameters })` → `tool({ inputSchema })` across 24 tool definitions in [web/src/app/api/chat/route.ts](web/src/app/api/chat/route.ts) + [web/src/app/api/meals/scan-chat/route.ts](web/src/app/api/meals/scan-chat/route.ts); (b) `maxSteps: N` → `stopWhen: stepCountIs(N)`; (c) `LanguageModelV1Middleware` → `LanguageModelV3Middleware` with `specificationVersion: "v3"` (import from `@ai-sdk/provider`); (d) `usage.promptTokens`/`completionTokens` → `inputTokens`/`outputTokens` in the `onStepFinish` token-budget tracker; (e) tool result property `.result` → `.output` in the onFinish step-pairing logic; (f) `StreamData` + `toDataStreamResponse({ data })` replaced with `toUIMessageStreamResponse({ messageMetadata })` — the #319 turn-complete sentinel now rides as metadata on the assistant message's finish part instead of a side-channel stream; (g) client-side `useChat` rewrite in [web/src/components/chat/chat-interface.tsx](web/src/components/chat/chat-interface.tsx) — v5 removed managed `input`/`handleInputChange`/`handleSubmit`/`isLoading`/`reload`, so the component now owns local input state, uses `DefaultChatTransport` with `prepareSendMessagesRequest` (flattens `UIMessage.parts` to the server's existing `{ role, content }` shape to avoid rewriting `/api/chat`'s message handling), and derives `isLoading` from `status`; (h) `Message` type replaced with `UIMessage` across 5 files — message content moved from `.content: string` to `.parts: Part[]`, `createdAt` moved into `message.metadata`; (i) `ToolStatusBar` updated for v5's per-tool UI part types (`tool-${name}` with `input-streaming`/`input-available`/`output-available`/`output-error` states); (j) `mimeType` → `mediaType` on `ImagePart` in [web/src/app/api/meals/analyze-photo/route.ts](web/src/app/api/meals/analyze-photo/route.ts). Three new-capability items unlocked by the upgrade are filed as follow-up issues rather than bundled into this PR: `generateObject` → `generateText({ output })` migration (4 call sites in meals routes, deprecated in v6 but functional), `stopWhen` / `ToolLoopAgent` refactor of the manual step-cap + synthesizer logic in the chat route, and Anthropic prompt caching on the system prompt + tool definitions.
 
 ### Fixed
