+++ web/src/app/api/chat/route.ts
@@ -1,73 +1,22 @@
 import { anthropic } from "@ai-sdk/anthropic";
 import { createGroq } from "@ai-sdk/groq";
-import { todayString, daysAgoString, startOfDayRFC3339, endOfDayRFC3339, addDays } from "@/lib/timezone";
-import { streamText, tool, jsonSchema, wrapLanguageModel, stepCountIs } from "ai";
+import { todayString } from "@/lib/timezone";
+import { streamText, wrapLanguageModel, stepCountIs } from "ai";
 import type { LanguageModelV3Middleware } from "@ai-sdk/provider";
 import { createServiceClient } from "@/lib/supabase/service";
 import { createClient } from "@/lib/supabase/server";
-import { syncSports } from "@/lib/sync/sports";
-import type { SportsCacheData } from "@/lib/sync/sports/provider";
-import { google } from "googleapis";
-import { getGoogleAuthClient } from "@/lib/google-auth";
-
-// Canonical result shape for state-mutating tools (issue #319).
-// Every mutating tool must return one of these so the model can reliably
-// distinguish "the action happened" from "the action didn't happen", and so
-// the synthesizer fallback never claims success for a failed tool result.
-type OkResult<T> = { ok: true } & T;
-type ErrResult = { ok: false; error: string };
-function ok<T extends Record<string, unknown>>(data: T): OkResult<T> {
-  return { ok: true, ...data };
-}
-function err(message: string): ErrResult {
-  return { ok: false, error: message };
-}
-
-// ── Demo mock data ───────────────────────────────────────────────────────────
-const DEMO_EMAILS = [
-  {
-    id: "demo-email-1",
-    from: "UPS Tracking <tracking@ups.com>",
-    subject: "Your package is out for delivery today",
-    date: "Mon, 13 Apr 2026 08:14:00 -0700",
-    body: "Hi Alex, your package is out for delivery today. Estimated delivery: today by 8pm. Track at ups.com.",
-  },
-  {
-    id: "demo-email-2",
-    from: "Alaska Airlines <noreply@alaskaair.com>",
-    subject: "Flight Confirmation: SFO → SEA Apr 20",
-    date: "Thu, 10 Apr 2026 11:02:00 -0700",
-    body: "Your flight AS 321 on April 20 from San Francisco (SFO) to Seattle (SEA) is confirmed. Departs 7:45am, arrives 9:50am. Confirmation code: KXZP94.",
-  },
-  {
-    id: "demo-email-3",
-    from: "Figma <notifications@figma.com>",
-    subject: "Action required: Accept team invite from Lena Park",
-    date: "Mon, 13 Apr 2026 09:47:00 -0700",
-    body: "Lena Park has invited you to join the 'Product Design' team on Figma. Click here to accept.",
-  },
-  {
-    id: "demo-email-4",
-    from: "Hacker News Digest <digest@hackernewsdigest.com>",
-    subject: "Top stories: Llama 4 benchmarks, SQLite as a backend",
-    date: "Mon, 13 Apr 2026 06:30:00 -0700",
-    body: "Top HN stories this week: Llama 4 beats GPT-4o on several benchmarks; SQLite as production backend — when it makes sense; Cloudflare Workers hits 100M active deployments.",
-  },
-  {
-    id: "demo-email-5",
-    from: "DocuSign <dse@docusign.net>",
-    subject: "Invoice #4421 — please sign",
-    date: "Fri, 11 Apr 2026 15:12:00 -0700",
-    body: "Alex Chen, a document has been sent to you for signature by Acme Corp. Invoice #4421 for $1,240.00 is ready for your review. This request will expire in 7 days.",
-  },
-];
-
-const DEMO_CALENDAR_EVENTS = [
-  { title: "Morning run", start: `${todayString()}T06:30:00`, end: `${todayString()}T07:15:00`, allDay: false, calendar: "Alex Chen", calendarType: "primary", location: null },
-  { title: "Team standup", start: `${todayString()}T09:00:00`, end: `${todayString()}T09:30:00`, allDay: false, calendar: "Alex Chen", calendarType: "primary", location: "Google Meet" },
-  { title: "Lunch w/ Priya", start: `${todayString()}T12:30:00`, end: `${todayString()}T13:30:00`, allDay: false, calendar: "Alex Chen", calendarType: "primary", location: "Tartine Manufactory" },
-  { title: "Gym — push day", start: `${todayString()}T18:00:00`, end: `${todayString()}T19:00:00`, allDay: false, calendar: "Alex Chen", calendarType: "primary", location: "Equinox SoMa" },
-];
+import type { ToolContext } from "@/lib/tools/_context";
+import { buildTasksTools } from "@/lib/tools/tasks";
+import { buildHabitsTools } from "@/lib/tools/habits";
+import { buildFitnessTools } from "@/lib/tools/fitness";
+import { buildProfileTools } from "@/lib/tools/profile";
+import { buildGmailTools } from "@/lib/tools/gmail";
+import { buildCalendarTools } from "@/lib/tools/calendar";
+import { buildMealsTools } from "@/lib/tools/meals";
+import { buildSessionTools } from "@/lib/tools/session";
+import { buildWorkoutTools } from "@/lib/tools/workouts";
+import { buildStocksTools } from "@/lib/tools/stocks";
+import { buildSportsTools } from "@/lib/tools/sports";
 
 const retryOnOverload: LanguageModelV3Middleware = {
   specificationVersion: "v3",
@@ -446,1346 +395,19 @@ ${userName ? `Address the user as "${userName}" — use their name naturally in 
     }))
     .filter((m: { role: "user" | "assistant"; content: string }) => m.content.trim() !== "");
 
-  function withTimeout<T>(promise: Promise<T>, ms: number, label: string): Promise<T> {
-    return Promise.race([
-      promise,
-      new Promise<never>((_, reject) =>
-        setTimeout(() => reject(new Error(`${label} timed out after ${ms}ms`)), ms)
-      ),
-    ]);
-  }
-
-  interface WorkoutExercise {
-    exercise: string;
-    sets?: number;
-    reps?: string;
-    weight_lbs?: number | null;
-    notes?: string | null;
-  }
-
-  function buildCalendarDescription(
-    warmup: WorkoutExercise[],
-    workout: WorkoutExercise[],
-    cooldown: WorkoutExercise[]
-  ): string {
-    const fmt = (ex: WorkoutExercise) => {
-      const parts = [ex.exercise];
-      if (ex.sets) parts.push(`${ex.sets} sets`);
-      if (ex.reps) parts.push(`× ${ex.reps}`);
-      if (ex.weight_lbs) parts.push(`@ ${ex.weight_lbs} lbs`);
-      return parts.join(" ");
-    };
-    const sections: string[] = [];
-    if (warmup.length) sections.push("Warm-up:\n" + warmup.map(fmt).join("\n"));
-    if (workout.length) sections.push("Workout:\n" + workout.map(fmt).join("\n"));
-    if (cooldown.length) sections.push("Cool-down:\n" + cooldown.map(fmt).join("\n"));
-    return sections.join("\n\n");
-  }
-
+  const toolContext: ToolContext = { supabase, userId, isDemo, sessionId };
   const tools = {
-    get_tasks: tool({
-      description: "Fetch tasks from the tasks table. Defaults to active tasks.",
-      inputSchema: jsonSchema<{ status?: "active" | "completed" | "archived" }>({
-        type: "object",
-        properties: {
-          status: {
-            type: "string",
-            enum: ["active", "completed", "archived"],
-            description: "Task status filter. Defaults to 'active'.",
-          },
-        },
-      }),
-      execute: async ({ status = "active" }) => {
-        let q = supabase
-          .from("tasks")
-          .select("id, title, priority, status, due_date, category, completed_at, created_at")
-          .eq("status", status)
-          .order("created_at", { ascending: false });
-        if (userId) q = q.eq("user_id", userId);
-        const { data, error } = await q;
-        if (error) return { error: error.message };
-        return data ?? [];
-      },
-    }),
-
-    add_task: tool({
-      description: "Add a new task or subtask to the tasks table. To add an item to a list (e.g. shopping list, grocery list), first call get_tasks to find the parent task ID, then call add_task with parent_id set.",
-      inputSchema: jsonSchema<{
-        title: string;
-        priority?: "high" | "medium" | "low";
-        category?: string;
-        due_date?: string;
-        parent_id?: string;
-      }>({
-        type: "object",
-        required: ["title"],
-        properties: {
-          title: { type: "string", description: "Task title." },
-          priority: {
-            type: "string",
-            enum: ["high", "medium", "low"],
-            description: "Task priority. Omit for subtasks.",
-          },
-          category: { type: "string", description: "Task category." },
-          due_date: { type: "string", description: "Due date in YYYY-MM-DD format. Omit for subtasks." },
-          parent_id: { type: "string", description: "Parent task UUID. Set this to add a subtask/list item under an existing task." },
-        },
-      }),
-      execute: async ({ title, priority, category, due_date, parent_id }) => {
-        if (due_date && !/^\d{4}-\d{2}-\d{2}$/.test(due_date)) {
-          return err(`due_date must be YYYY-MM-DD format, got: "${due_date}"`);
-        }
-
-        // Deduplication guard — prevents double-inserts from stream retries
-        const windowStart = new Date(Date.now() - 90_000).toISOString();
-        let dupQuery = supabase
-          .from("tasks")
-          .select("id, title, priority, status, due_date, category, parent_id, created_at")
-          .eq("user_id", userId)
-          .eq("status", "active")
-          .ilike("title", title.trim())
-          .gte("created_at", windowStart);
-
-        if (due_date) {
-          dupQuery = dupQuery.eq("due_date", due_date);
-        } else {
-          dupQuery = dupQuery.is("due_date", null);
-        }
-
-        const { data: existing } = await dupQuery.maybeSingle();
-        if (existing) return ok({ task: existing, deduped: true });
-
-        const { data, error: insertError } = await supabase
-          .from("tasks")
-          .insert({
-            user_id: userId,
-            title,
-            priority: parent_id ? null : (priority ?? null),
-            category: category ?? null,
-            due_date: parent_id ? null : (due_date ?? null),
-            status: "active",
-            parent_id: parent_id ?? null,
-          })
-          .select("id, title, priority, status, due_date, category, parent_id, created_at")
-          .single();
-        if (insertError) return err(insertError.message);
-        if (!data) return err("Insert returned no row — task may not have been saved.");
-        return ok({ task: data });
-      },
-    }),
-
-    complete_task: tool({
-      description: "Mark a task as completed by its ID.",
-      inputSchema: jsonSchema<{ id: string }>({
-        type: "object",
-        required: ["id"],
-        properties: {
-          id: { type: "string", description: "Task UUID." },
-        },
-      }),
-      execute: async ({ id }) => {
-        let q = supabase
-          .from("tasks")
-          .update({ status: "completed", completed_at: new Date().toISOString() })
-          .eq("id", id);
-        if (userId) q = q.eq("user_id", userId);
-        const { data, error: updateError } = await q.select("id, title, status, completed_at").maybeSingle();
-        if (updateError) return err(updateError.message);
-        if (!data) return err(`No active task found with id ${id} for this user — nothing was changed.`);
-        return ok({ task: data });
-      },
-    }),
-
-    get_habits_today: tool({
-      description: "Get all active habits and their completion status for today (or a specified date).",
-      inputSchema: jsonSchema<{ date?: string }>({
-        type: "object",
-        properties: {
-          date: { type: "string", description: "Date in YYYY-MM-DD format. Defaults to today." },
-        },
-      }),
-      execute: async ({ date }) => {
-        const targetDate = date ?? todayString();
-        let regQ = supabase.from("habit_registry").select("id, name, emoji, category").eq("active", true);
-        let logQ = supabase.from("habits").select("habit_id, completed, notes").eq("date", targetDate);
-        if (userId) { regQ = regQ.eq("user_id", userId); logQ = logQ.eq("user_id", userId); }
-        const [registryResult, logsResult] = await Promise.all([regQ, logQ]);
-        if (registryResult.error) return { error: registryResult.error.message };
-        const logMap = new Map(
-          (logsResult.data ?? []).map((l: { habit_id: string; completed: boolean; notes: string | null }) => [l.habit_id, l])
-        );
-        return (registryResult.data ?? []).map((h: { id: string; name: string; emoji: string | null; category: string | null }) => ({
-          id: h.id,
-          name: h.name,
-          emoji: h.emoji,
-          category: h.category,
-          completed: logMap.get(h.id)?.completed ?? false,
-          notes: logMap.get(h.id)?.notes ?? null,
-        }));
-      },
-    }),
-
-    log_habit: tool({
-      description: "Log a habit as completed. Looks up the habit by name from the registry.",
-      inputSchema: jsonSchema<{ name: string; date?: string; notes?: string }>({
-        type: "object",
-        required: ["name"],
-        properties: {
-          name: { type: "string", description: "Habit name (case-insensitive partial match)." },
-          date: { type: "string", description: "Date in YYYY-MM-DD format. Defaults to today." },
-          notes: { type: "string", description: "Optional notes." },
-        },
-      }),
-      execute: async ({ name, date, notes }) => {
-        const targetDate = date ?? todayString();
-        let habQ = supabase
-          .from("habit_registry")
-          .select("id, name")
-          .ilike("name", `%${name}%`)
-          .eq("active", true)
-          .limit(1);
-        if (userId) habQ = habQ.eq("user_id", userId);
-        const { data: habits, error: lookupError } = await habQ;
-        if (lookupError) return err(lookupError.message);
-        if (!habits || habits.length === 0) return err(`No active habit matching "${name}" found.`);
-        const habit = habits[0];
-        const { data, error: upsertError } = await supabase
-          .from("habits")
-          .upsert(
-            { user_id: userId, habit_id: habit.id, date: targetDate, completed: true, notes: notes ?? null },
-            { onConflict: "habit_id,date" }
-          )
-          .select("habit_id, date, completed, notes")
-          .single();
-        if (upsertError) return err(upsertError.message);
-        if (!data) return err("Upsert returned no row — habit may not have been logged.");
-        return ok({ habit: habit.name, log: data });
-      },
-    }),
-
-    get_fitness_summary: tool({
-      description: "Get recent body composition, workouts, and recovery metrics.",
-      inputSchema: jsonSchema<{ days?: number }>({
-        type: "object",
-        properties: {
-          days: {
-            type: "number",
-            description: "Number of days back to include for workouts. Defaults to 7.",
-          },
-        },
-      }),
-      execute: async ({ days = 7 }) => {
-        const sinceStr = daysAgoString(days);
-
-        let bodyQ = supabase
-          .from("fitness_log")
-          .select("date, weight_lb, body_fat_pct, bmi, muscle_mass_lb, visceral_fat, source")
-          .not("body_fat_pct", "is", null)
-          .order("date", { ascending: false })
-          .limit(2);
-        let workQ = supabase
-          .from("workout_sessions")
-          .select("date, activity, duration_mins, calories, avg_hr, notes")
-          .gte("date", sinceStr)
-          .order("date", { ascending: false });
-        let recQ = supabase
-          .from("recovery_metrics")
-          .select("date, avg_hrv, resting_hr, sleep_score, readiness, source")
-          .order("date", { ascending: false })
-          .limit(1);
-        if (userId) { bodyQ = bodyQ.eq("user_id", userId); workQ = workQ.eq("user_id", userId); recQ = recQ.eq("user_id", userId); }
-
-        const [bodyCompResult, workoutsResult, recoveryResult] = await Promise.all([bodyQ, workQ, recQ]);
-
-        return {
-          body_composition: bodyCompResult.data ?? [],
-          workouts: workoutsResult.data ?? [],
-          recovery: recoveryResult.data?.[0] ?? null,
-        };
-      },
-    }),
-
-    get_profile: tool({
-      description: "Get all profile key/value entries.",
-      inputSchema: jsonSchema<Record<string, never>>({
-        type: "object",
-        properties: {},
-      }),
-      execute: async () => {
-        let q = supabase.from("profile").select("key, value, updated_at").order("key", { ascending: true });
-        if (userId) q = q.eq("user_id", userId);
-        const { data, error } = await q;
-        if (error) return { error: error.message };
-        return data ?? [];
-      },
-    }),
-
-    update_profile: tool({
-      description:
-        "Upsert one or more profile key/value pairs. Use when the user explicitly agrees to save a goal, preference, or personal target. " +
-        "For known fitness/nutrition goals use the canonical flat keys (weight_goal_lbs, body_fat_goal_pct, weekly_workout_goal, " +
-        "weekly_active_cal_goal, calorie_goal, protein_goal, carbs_goal, fat_goal, fiber_goal) so they surface in the web UI. " +
-        "For other goals use dot-notation (sleep.goal.hrs, study.goal.mins_per_day, etc.). " +
-        "Always tell the user what you are about to write before calling this tool, then confirm each key that was saved.",
-      inputSchema: jsonSchema<{ updates: { key: string; value: string }[] }>({
-        type: "object",
-        required: ["updates"],
-        properties: {
-          updates: {
-            type: "array",
-            description: "Key/value pairs to upsert into the profile table.",
-            items: {
-              type: "object",
-              required: ["key", "value"],
-              properties: {
-                key: { type: "string", description: "Profile key." },
-                value: { type: "string", description: "Value to store (always a string)." },
-              },
-            },
-          },
-        },
-      }),
-      execute: async ({ updates }) => {
-        const rows = updates.map(({ key, value }) => ({ key, value, ...(userId ? { user_id: userId } : {}) }));
-        const { data, error: upsertError } = await supabase
-          .from("profile")
-          .upsert(rows, { onConflict: "user_id,key" })
-          .select("key, value, updated_at");
-        if (upsertError) return err(upsertError.message);
-        if (!data || data.length !== rows.length) {
-          return err(`Profile upsert returned ${data?.length ?? 0} rows, expected ${rows.length} — values may not be saved.`);
-        }
-        return ok({ written: data });
-      },
-    }),
-
-    search_gmail: tool({
-      description:
-        "Search Gmail using any Gmail query string (e.g. 'from:regal tickets', 'subject:invoice is:unread'). Returns message ID, sender, subject, and date. Use get_email_body to read the full content of a specific message.",
-      inputSchema: jsonSchema<{ query: string; max_results?: number }>({
-        type: "object",
-        required: ["query"],
-        properties: {
-          query: {
-            type: "string",
-            description: "Gmail search query. Supports all Gmail search operators.",
-          },
-          max_results: {
-            type: "number",
-            description: "Max messages to return. Defaults to 5, max 10.",
-          },
-        },
-      }),
-      execute: async ({ query, max_results = 5 }) => {
-        if (isDemo) {
-          // Return mock emails filtered loosely by query keywords
-          const q = query.toLowerCase();
-          const mockEmails = DEMO_EMAILS.filter((e) =>
-            !q || e.subject.toLowerCase().includes(q.split(" ")[0]) || q.includes("unread") || q.includes("subject:")
-          ).slice(0, Math.min(max_results, 5));
-          return { results: mockEmails };
-        }
-        try {
-          const auth = getGoogleAuthClient();
-          const gmail = google.gmail({ version: "v1", auth });
-
-          const listRes = await gmail.users.messages.list({
-            userId: "me",
-            q: query,
-            maxResults: Math.min(max_results, 10),
-          });
-
-          const messages = listRes.data.messages ?? [];
-          if (messages.length === 0) return { results: [] };
-
-          const getHeader = (
-            headers: { name?: string | null; value?: string | null }[],
-            name: string
-          ) => headers.find((h) => h.name?.toLowerCase() === name.toLowerCase())?.value ?? "";
-
-          const summaries = await Promise.all(
-            messages.map(async (m) => {
-              const msg = await gmail.users.messages.get({
-                userId: "me",
-                id: m.id!,
-                format: "metadata",
-                metadataHeaders: ["From", "Subject", "Date"],
-              });
-              const headers = msg.data.payload?.headers ?? [];
-              return {
-                id: m.id,
-                from: getHeader(headers, "From"),
-                subject: getHeader(headers, "Subject") || "(No subject)",
-                date: getHeader(headers, "Date"),
-              };
-            })
-          );
-
-          return { results: summaries };
-        } catch (err) {
-          return { error: err instanceof Error ? err.message : "Gmail search failed" };
-        }
-      },
-    }),
-
-    get_email_body: tool({
-      description:
-        "Fetch the full text body of a Gmail message by its ID. Use this after search_gmail to read email content. Returns decoded plain text.",
-      inputSchema: jsonSchema<{ message_id: string }>({
-        type: "object",
-        required: ["message_id"],
-        properties: {
-          message_id: {
-            type: "string",
-            description: "The Gmail message ID from search_gmail results.",
-          },
-        },
-      }),
-      execute: async ({ message_id }) => {
-        if (isDemo) {
-          const email = DEMO_EMAILS.find((e) => e.id === message_id);
-          if (!email) return { error: "Email not found" };
-          return { id: message_id, from: email.from, subject: email.subject, date: email.date, body: email.body ?? email.subject, truncated: false };
-        }
-        try {
-          const auth = getGoogleAuthClient();
-          const gmail = google.gmail({ version: "v1", auth });
-
-          const msg = await gmail.users.messages.get({
-            userId: "me",
-            id: message_id,
-            format: "full",
-          });
-
-          const headers = msg.data.payload?.headers ?? [];
-          const getHeader = (name: string) =>
-            headers.find((h) => h.name?.toLowerCase() === name.toLowerCase())?.value ?? "";
-
-          type Part = {
-            mimeType?: string | null;
-            body?: { data?: string | null } | null;
-            parts?: Part[] | null;
-          };
-
-          function findBody(parts: Part[], mime: string): string | null {
-            for (const part of parts) {
-              if (part.mimeType === mime && part.body?.data) {
-                return Buffer.from(part.body.data, "base64url").toString("utf-8");
-              }
-              if (part.parts) {
-                const found = findBody(part.parts, mime);
-                if (found) return found;
-              }
-            }
-            return null;
-          }
-
-          const payload = msg.data.payload;
-          let body: string | null = null;
-
-          if (payload?.body?.data) {
-            body = Buffer.from(payload.body.data, "base64url").toString("utf-8");
-          } else if (payload?.parts) {
-            body = findBody(payload.parts as Part[], "text/plain");
-            if (!body) {
-              const html = findBody(payload.parts as Part[], "text/html");
-              if (html) body = html.replace(/<[^>]+>/g, " ").replace(/\s+/g, " ").trim();
-            }
-          }
-
-          const isTruncated = body ? body.length > 4000 : false;
-          const bodyText = body
-            ? isTruncated
-              ? body.slice(0, 4000) + `\n\n[...email truncated — ${body.length - 4000} more characters not shown]`
-              : body
-            : null;
-
-          return {
-            id: message_id,
-            from: getHeader("From"),
-            subject: getHeader("Subject"),
-            date: getHeader("Date"),
-            body: bodyText ?? "(No readable body found)",
-            truncated: isTruncated,
-          };
-        } catch (err) {
-          return { error: err instanceof Error ? err.message : "Failed to fetch email body" };
-        }
-      },
-    }),
-
-    list_calendar_events: tool({
-      description:
-        "List events across all Google Calendars for a given date range. Defaults to today only. Use this whenever the user asks what's on their calendar, schedule, or agenda. Each event in the result includes an eventId field — preserve and use it when calling delete_calendar_event or update_calendar_event.",
-      inputSchema: jsonSchema<{ date?: string; days?: number }>({
-        type: "object",
-        properties: {
-          date: {
-            type: "string",
-            description: "Start date in YYYY-MM-DD format. Defaults to today.",
-          },
-          days: {
-            type: "number",
-            description: "Number of days to include (1 = just the start date). Defaults to 1.",
-          },
-        },
-      }),
-      execute: async ({ date, days = 1 }) => {
-        if (isDemo) {
-          return { events: DEMO_CALENDAR_EVENTS, count: DEMO_CALENDAR_EVENTS.length };
-        }
-        try {
-          const startDate = date ?? todayString();
-          const endDate = addDays(startDate, Math.max(1, days) - 1);
-
-          const auth = getGoogleAuthClient();
-          const calendar = google.calendar({ version: "v3", auth });
-
-          const calListRes = await calendar.calendarList.list({ minAccessRole: "reader" });
-          const calendars = calListRes.data.items ?? [];
-
-          const allEventArrays = await Promise.all(
-            calendars.map(async (cal) => {
-              const res = await calendar.events.list({
-                calendarId: cal.id!,
-                timeMin: startOfDayRFC3339(startDate),
-                timeMax: endOfDayRFC3339(endDate),
-                singleEvents: true,
-                orderBy: "startTime",
-                maxResults: 25,
-              });
-              const calName = cal.summaryOverride ?? cal.summary ?? cal.id ?? "Unknown";
-              const calNameLower = calName.toLowerCase();
-              const calendarType = cal.primary
-                ? "primary"
-                : calNameLower.includes("birthday") || calNameLower.includes("contact")
-                ? "birthday"
-                : calNameLower.includes("holiday")
-                ? "holiday"
-                : "other";
-
-              return (res.data.items ?? [])
-                .filter((e) => {
-                  if (e.status === "cancelled") return false;
-                  // Filter out events the user has declined
-                  const selfAttendee = e.attendees?.find((a) => a.self);
-                  if (selfAttendee?.responseStatus === "declined") return false;
-                  return true;
-                })
-                .map((e) => ({
-                  eventId: e.id ?? "",
-                  title: e.summary ?? "(No title)",
-                  start: e.start?.dateTime ?? e.start?.date ?? "",
-                  end: e.end?.dateTime ?? e.end?.date ?? "",
-                  allDay: !e.start?.dateTime,
-                  calendar: calName,
-                  calendarType,
-                  location: e.location ?? null,
-                }));
-            })
-          );
-
-          const events = allEventArrays
-            .flat()
-            .sort((a, b) => a.start.localeCompare(b.start));
-
-          return { events, count: events.length };
-        } catch (err) {
-          return { error: err instanceof Error ? err.message : "Failed to list calendar events" };
-        }
-      },
-    }),
-
-    create_calendar_event: tool({
-      description:
-        "Create a new event on the primary Google Calendar. For timed events, provide date + start_time. For all-day events, set all_day: true and provide only date.",
-      inputSchema: jsonSchema<{
-        title: string;
-        date: string;
-        start_time?: string;
-        end_time?: string;
-        location?: string;
-        description?: string;
-        all_day?: boolean;
-      }>({
-        type: "object",
-        required: ["title", "date"],
-        properties: {
-          title: { type: "string", description: "Event title/summary." },
-          date: { type: "string", description: "Date in YYYY-MM-DD format." },
-          start_time: {
-            type: "string",
-            description: "Start time in HH:MM (24h) format. Required for timed events.",
-          },
-          end_time: {
-            type: "string",
-            description: "End time in HH:MM (24h) format. Defaults to start_time + 2 hours.",
-          },
-          location: { type: "string", description: "Event location." },
-          description: { type: "string", description: "Event description or notes." },
-          all_day: {
-            type: "boolean",
-            description: "If true, creates an all-day event. start_time and end_time are ignored.",
-          },
-        },
-      }),
-      execute: async ({ title, date, start_time, end_time, location, description, all_day = false }) => {
-        if (isDemo) {
-          return ok({
-            id: `demo-${Date.now()}`,
-            title,
-            start: { date, dateTime: start_time ? `${date}T${start_time}:00` : date },
-            end: { date, dateTime: end_time ? `${date}T${end_time}:00` : date },
-            note: "Demo mode — event not saved to real calendar.",
-          });
-        }
-        try {
-          if (!/^\d{4}-\d{2}-\d{2}$/.test(date)) {
-            return err(`date must be YYYY-MM-DD, got: "${date}"`);
-          }
-          if (!all_day && !start_time) {
-            return err("start_time is required for timed events. Use all_day: true for all-day events.");
-          }
-
-          const auth = getGoogleAuthClient();
-          const calendar = google.calendar({ version: "v3", auth });
-          const tz = process.env.USER_TIMEZONE ?? "America/Los_Angeles";
-
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          let eventBody: Record<string, any>;
-
-          if (all_day) {
-            eventBody = {
-              summary: title,
-              start: { date },
-              end: { date },
-              ...(location ? { location } : {}),
-              ...(description ? { description } : {}),
-            };
-          } else {
-            let computedEnd = end_time;
-            if (!computedEnd) {
-              const [h, m] = start_time!.split(":").map(Number);
-              computedEnd = `${String((h + 2) % 24).padStart(2, "0")}:${String(m).padStart(2, "0")}`;
-            }
-            eventBody = {
-              summary: title,
-              start: { dateTime: `${date}T${start_time}:00`, timeZone: tz },
-              end: { dateTime: `${date}T${computedEnd}:00`, timeZone: tz },
-              ...(location ? { location } : {}),
-              ...(description ? { description } : {}),
-            };
-          }
-
-          const res = await calendar.events.insert({
-            calendarId: "primary",
-            requestBody: eventBody,
-          });
-          if (!res.data.id) return err("Google returned no event id — create may not have persisted.");
-
-          // Read-after-write verification (#319): confirm the event Google
-          // actually stored matches what we asked for. Catches wrong-target /
-          // partial-write / silent-rewrite classes that a 200 response can't.
-          const verify = await calendar.events.get({ calendarId: "primary", eventId: res.data.id });
-          const verifiedStart = verify.data.start?.dateTime ?? verify.data.start?.date ?? null;
-          const verifiedEnd = verify.data.end?.dateTime ?? verify.data.end?.date ?? null;
-          const expectedStart = all_day ? date : `${date}T${start_time}:00`;
-          if (!verifiedStart?.startsWith(expectedStart.slice(0, all_day ? 10 : 16))) {
-            return err(
-              `Calendar accepted the create but a follow-up read shows start=${verifiedStart}, expected to start with ${expectedStart}. The event may not be where you wanted it.`
-            );
-          }
-
-          return ok({
-            id: res.data.id,
-            title: res.data.summary,
-            start: res.data.start,
-            end: res.data.end,
-            link: res.data.htmlLink,
-            verified: { start: verifiedStart, end: verifiedEnd },
-          });
-        } catch (caughtErr) {
-          return err(caughtErr instanceof Error ? caughtErr.message : "Failed to create calendar event");
-        }
-      },
-    }),
-
-    delete_calendar_event: tool({
-      description:
-        "Delete a Google Calendar event by eventId. IMPORTANT: before calling this tool, always state the event title, date, and time to the user and require explicit confirmation. Never delete without confirmed user intent.",
-      inputSchema: jsonSchema<{ eventId: string }>({
-        type: "object",
-        required: ["eventId"],
-        properties: {
-          eventId: { type: "string", description: "The eventId from list_calendar_events." },
-        },
-      }),
-      execute: async ({ eventId }) => {
-        if (isDemo) {
-          return ok({ eventId, note: "Demo mode — event not deleted from real calendar." });
-        }
-        try {
-          const auth = getGoogleAuthClient();
-          const calendar = google.calendar({ version: "v3", auth });
-          await calendar.events.delete({ calendarId: "primary", eventId });
-
-          // Read-after-write verification (#319): confirm the event is gone or
-          // marked cancelled. Google delete returns 204 even if the event was
-          // already gone, so we re-read to be sure.
-          try {
-            const verify = await calendar.events.get({ calendarId: "primary", eventId });
-            if (verify.data.status !== "cancelled") {
-              return err(`Calendar accepted the delete but a follow-up read still shows status=${verify.data.status}. The event may not be deleted.`);
-            }
-          } catch (verifyErr) {
-            // 404/410 on get-after-delete means the event is gone — that's success.
-            const status = (verifyErr as { code?: number; response?: { status?: number } }).code
-              ?? (verifyErr as { response?: { status?: number } }).response?.status;
-            if (status !== 404 && status !== 410) {
-              return err(
-                `Calendar accepted the delete but verification failed: ${verifyErr instanceof Error ? verifyErr.message : "unknown error"}`
-              );
-            }
-          }
-
-          return ok({ eventId });
-        } catch (caughtErr) {
-          return err(caughtErr instanceof Error ? caughtErr.message : "Failed to delete calendar event");
-        }
-      },
-    }),
-
-    update_calendar_event: tool({
-      description:
-        "Update an existing Google Calendar event by eventId. Accepts any subset of: summary (title), start (RFC3339 dateTime or date), end (RFC3339 dateTime or date), location, description. IMPORTANT: before calling, state the before/after changes and require explicit user confirmation.",
-      inputSchema: jsonSchema<{
-        eventId: string;
-        summary?: string;
-        start?: string;
-        end?: string;
-        location?: string;
-        description?: string;
-      }>({
-        type: "object",
-        required: ["eventId"],
-        properties: {
-          eventId: { type: "string", description: "The eventId from list_calendar_events." },
-          summary: { type: "string", description: "New event title." },
-          start: { type: "string", description: "New start as RFC3339 dateTime (e.g. 2026-04-15T09:00:00) or date (YYYY-MM-DD) for all-day." },
-          end: { type: "string", description: "New end as RFC3339 dateTime or date." },
-          location: { type: "string", description: "New location string." },
-          description: { type: "string", description: "New description/notes." },
-        },
-      }),
-      execute: async ({ eventId, summary, start, end, location, description }) => {
-        if (isDemo) {
-          return ok({ eventId, note: "Demo mode — event not updated in real calendar." });
-        }
-        try {
-          const auth = getGoogleAuthClient();
-          const calendar = google.calendar({ version: "v3", auth });
-          const tz = process.env.USER_TIMEZONE ?? "America/Los_Angeles";
-
-          // Build patch body — only include fields that were provided
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          const requestBody: Record<string, any> = {};
-          if (summary !== undefined) requestBody.summary = summary;
-          if (location !== undefined) requestBody.location = location;
-          if (description !== undefined) requestBody.description = description;
-          if (start !== undefined) {
-            requestBody.start = /T/.test(start)
-              ? { dateTime: start.length === 16 ? `${start}:00` : start, timeZone: tz }
-              : { date: start };
-          }
-          if (end !== undefined) {
-            requestBody.end = /T/.test(end)
-              ? { dateTime: end.length === 16 ? `${end}:00` : end, timeZone: tz }
-              : { date: end };
-          }
-
-          const res = await calendar.events.patch({
-            calendarId: "primary",
-            eventId,
-            requestBody,
-          });
-
-          // Read-after-write verification (#319): the original symptom of the
-          // bug was a "successful" patch that didn't actually change the event.
-          // Re-read and check the requested fields match what we asked for.
-          const verify = await calendar.events.get({ calendarId: "primary", eventId });
-          const mismatches: string[] = [];
-          if (summary !== undefined && verify.data.summary !== summary) {
-            mismatches.push(`summary expected "${summary}", got "${verify.data.summary}"`);
-          }
-          if (start !== undefined) {
-            const verifiedStart = verify.data.start?.dateTime ?? verify.data.start?.date ?? "";
-            const expectedPrefix = /T/.test(start) ? start.slice(0, 16) : start;
-            if (!verifiedStart.startsWith(expectedPrefix)) {
-              mismatches.push(`start expected to begin with "${expectedPrefix}", got "${verifiedStart}"`);
-            }
-          }
-          if (end !== undefined) {
-            const verifiedEnd = verify.data.end?.dateTime ?? verify.data.end?.date ?? "";
-            const expectedPrefix = /T/.test(end) ? end.slice(0, 16) : end;
-            if (!verifiedEnd.startsWith(expectedPrefix)) {
-              mismatches.push(`end expected to begin with "${expectedPrefix}", got "${verifiedEnd}"`);
-            }
-          }
-          if (location !== undefined && verify.data.location !== location) {
-            mismatches.push(`location expected "${location}", got "${verify.data.location}"`);
-          }
-          if (mismatches.length > 0) {
-            return err(`Calendar accepted the update but a follow-up read shows it didn't take: ${mismatches.join("; ")}`);
-          }
-
-          return ok({
-            eventId: res.data.id,
-            title: res.data.summary,
-            start: res.data.start,
-            end: res.data.end,
-            link: res.data.htmlLink,
-            verified: true,
-          });
-        } catch (caughtErr) {
-          return err(caughtErr instanceof Error ? caughtErr.message : "Failed to update calendar event");
-        }
-      },
-    }),
-
-    get_recipes: tool({
-      description: "Search saved recipes by ingredient, name, or tag. Omit query to return all recipes.",
-      inputSchema: jsonSchema<{ query?: string }>({
-        type: "object",
-        properties: {
-          query: { type: "string", description: "Ingredient, recipe name, or tag to search for." },
-        },
-      }),
-      execute: async ({ query }) => {
-        let req = supabase
-          .from("recipes")
-          .select("id, name, cuisine, ingredients, instructions, tags");
-        if (userId) req = req.eq("user_id", userId);
-        if (query) {
-          req = req.or(`name.ilike.%${query}%,ingredients.ilike.%${query}%`);
-        }
-        const { data, error } = await req.order("name");
-        if (error) return { error: error.message };
-        return data ?? [];
-      },
-    }),
-
-    get_today_meals: tool({
-      description: "Get all meals logged today. Call this before making any claim about what the user has or hasn't eaten today.",
-      inputSchema: jsonSchema<Record<string, never>>({
-        type: "object",
-        properties: {},
-      }),
-      execute: async () => {
-        const today = todayString();
-        let q = supabase
-          .from("meal_log")
-          .select("meal_type, notes, calories, protein_g, carbs_g, fat_g, recipes(name)")
-          .eq("date", today);
-        if (userId) q = q.eq("user_id", userId);
-        const { data } = await q;
-        return data ?? [];
-      },
-    }),
-
-    get_session_history: tool({
-      description: "Fetch earlier messages from this chat session. Use when the user references something said earlier that isn't in the current context window. Always ask the user before calling: \"Should I pull earlier messages from this session for more context?\"",
-      inputSchema: jsonSchema<{ limit?: number }>({
-        type: "object",
-        properties: {
-          limit: { type: "number", description: "How many messages to fetch (max 40). Defaults to 20." },
-        },
-      }),
-      execute: async ({ limit = 20 }) => {
-        if (!sessionId) return { error: "No session ID available." };
-        const { data } = await supabase
-          .from("chat_messages")
-          .select("role, content, created_at")
-          .eq("session_id", sessionId)
-          .in("role", ["user", "assistant"])
-          .order("position", { ascending: true, nullsFirst: false })
-          .limit(Math.min(limit, 40));
-        return data ?? [];
-      },
-    }),
-
-    get_workout_plan: tool({
-      description: "Fetch the workout plan for the current Mon–Sun week. No parameters. Call this before making any adjustment or suggestion to the user's workout program.",
-      inputSchema: jsonSchema<Record<string, never>>({
-        type: "object",
-        properties: {},
-      }),
-      execute: async () => {
-        if (!userId) return { error: "Not authenticated" };
-        if (isDemo) return { demo: true, note: "Demo mode — no real workout plans." };
-        const todayStr = todayString();
-        const dow = (new Date(`${todayStr}T12:00:00Z`).getUTCDay() + 6) % 7; // 0=Mon, 6=Sun
-        const monday = addDays(todayStr, -dow);
-        const sunday = addDays(monday, 6);
-        const { data, error } = await supabase
-          .from("workout_plans")
-          .select("*")
-          .eq("user_id", userId)
-          .gte("date", monday)
-          .lte("date", sunday)
-          .order("date", { ascending: true });
-        if (error) return { error: error.message };
-        return { week: `${monday} – ${sunday}`, plans: data ?? [] };
-      },
-    }),
-
-    assign_workout: tool({
-      description: "Assign or replace one day's workout plan. Upserts to workout_plans and optionally creates/updates the matching Google Calendar event. Pre-flight required when update_calendar is true: (1) ask the user what time their workout will be if start_time is unknown, (2) call list_calendar_events for the target date, check for time overlaps and duplicate Workout titles, surface any conflicts to the user, and get confirmation before calling this tool.",
-      inputSchema: jsonSchema<{
-        date: string;
-        name?: string;
-        warmup: WorkoutExercise[];
-        workout: WorkoutExercise[];
-        cooldown: WorkoutExercise[];
-        notes?: string;
-        start_time?: string;
-        end_time?: string;
-        update_calendar?: boolean;
-      }>({
-        type: "object",
-        required: ["date", "warmup", "workout", "cooldown"],
-        properties: {
-          date: { type: "string", description: "Date in YYYY-MM-DD format." },
-          name: { type: "string", description: "Workout name shown on the fitness page card, e.g. 'Push Day', 'Pull Day', 'Legs'." },
-          warmup: { type: "array", items: { type: "object" }, description: "Warm-up exercises." },
-          workout: { type: "array", items: { type: "object" }, description: "Main workout exercises." },
-          cooldown: { type: "array", items: { type: "object" }, description: "Cool-down exercises." },
-          notes: { type: "string", description: "Optional plan notes." },
-          start_time: { type: "string", description: "Workout start time in HH:MM (24h) format." },
-          end_time: { type: "string", description: "Workout end time in HH:MM (24h) format. Defaults to start_time + 1 hour." },
-          update_calendar: { type: "boolean", description: "Create/update a Google Calendar event. Default true." },
-        },
-      }),
-      execute: async ({ date, name, warmup, workout, cooldown, notes, start_time, end_time, update_calendar = true }) => {
-        if (!userId) return err("Not authenticated");
-        if (isDemo) return ok({ demo: true, note: "Demo mode — plan not saved." });
-
-        const { data: upserted, error: upsertError } = await supabase
-          .from("workout_plans")
-          .upsert(
-            { user_id: userId, date, name: name ?? null, warmup, workout, cooldown, notes: notes ?? null, updated_at: new Date().toISOString() },
-            { onConflict: "user_id,date" }
-          )
-          .select()
-          .single();
-        if (upsertError) return err(upsertError.message);
-        if (!upserted) return err("Workout plan upsert returned no row — plan may not have been saved.");
-
-        if (!update_calendar) {
-          return ok({ plan: upserted, calendar_synced: false });
-        }
-
-        try {
-          const auth = getGoogleAuthClient();
-          const calendar = google.calendar({ version: "v3", auth });
-          const description = buildCalendarDescription(warmup, workout, cooldown);
-          const title = `Workout — ${new Date(date + "T00:00:00").toLocaleDateString("en-US", { weekday: "long" })}`;
-          const tz = process.env.USER_TIMEZONE ?? "America/Los_Angeles";
-
-          let startObj: object;
-          let endObj: object;
-          if (start_time) {
-            const computedEnd = end_time ?? (() => {
-              const [h, m] = start_time.split(":").map(Number);
-              return `${String((h + 1) % 24).padStart(2, "0")}:${String(m).padStart(2, "0")}`;
-            })();
-            startObj = { dateTime: `${date}T${start_time}:00`, timeZone: tz };
-            endObj = { dateTime: `${date}T${computedEnd}:00`, timeZone: tz };
-          } else {
-            startObj = { date };
-            endObj = { date };
-          }
-
-          let eventId: string | null = upserted.calendar_event_id ?? null;
-          if (eventId) {
-            await withTimeout(
-              calendar.events.patch({
-                calendarId: "primary",
-                eventId,
-                requestBody: { summary: title, description, start: startObj, end: endObj },
-              }),
-              8000,
-              "calendar patch"
-            );
-          } else {
-            const res = await withTimeout(
-              calendar.events.insert({
-                calendarId: "primary",
-                requestBody: { summary: title, start: startObj, end: endObj, description },
-              }),
-              8000,
-              "calendar insert"
-            );
-            eventId = res.data.id ?? null;
-            await supabase
-              .from("workout_plans")
-              .update({ calendar_event_id: eventId })
-              .eq("user_id", userId)
-              .eq("date", date);
-            upserted.calendar_event_id = eventId;
-          }
-          return ok({ plan: upserted, calendar_synced: true });
-        } catch (calErr) {
-          // Plan saved but calendar sync failed — partial success surfaced to
-          // the model so it tells the user instead of claiming full success.
-          return ok({
-            plan: upserted,
-            calendar_synced: false,
-            calendar_error: calErr instanceof Error ? calErr.message : "Calendar sync failed",
-          });
-        }
-      },
-    }),
-
-    update_workout_exercise: tool({
-      description: "Patch a single exercise in one phase of an existing workout plan. Fetches the row, finds the exercise by name (case-insensitive), merges the updates, upserts back, and refreshes the calendar event description. Supports renaming/swapping the exercise via updates.exercise.",
-      inputSchema: jsonSchema<{
-        date: string;
-        phase: "warmup" | "workout" | "cooldown";
-        exercise_name: string;
-        updates: { exercise?: string; sets?: number; reps?: string; weight_lbs?: number; notes?: string };
-      }>({
-        type: "object",
-        required: ["date", "phase", "exercise_name", "updates"],
-        properties: {
-          date: { type: "string", description: "YYYY-MM-DD date of the plan to edit." },
-          phase: { type: "string", enum: ["warmup", "workout", "cooldown"], description: "Which phase the exercise is in." },
-          exercise_name: { type: "string", description: "Existing exercise name to match (case-insensitive)." },
-          updates: {
-            type: "object",
-            description: "Fields to merge into the matched exercise object. Use `exercise` to rename/swap.",
-            properties: {
-              exercise: { type: "string", description: "New exercise name (for swaps/renames)." },
-              sets: { type: "number" },
-              reps: { type: "string" },
-              weight_lbs: { type: "number" },
-              notes: { type: "string" },
-            },
-          },
-        },
-      }),
-      execute: async ({ date, phase, exercise_name, updates }) => {
-        if (!userId) return err("Not authenticated");
-        if (isDemo) return ok({ demo: true, note: "Demo mode — not saved." });
-
-        const { data: row, error: fetchErr } = await supabase
-          .from("workout_plans")
-          .select("*")
-          .eq("user_id", userId)
-          .eq("date", date)
-          .single();
-        if (fetchErr) return err(fetchErr.message);
-        if (!row) return err("No plan found for that date.");
-
-        const arr: WorkoutExercise[] = [...(row[phase] as WorkoutExercise[])];
-        const idx = arr.findIndex((e) => e.exercise.toLowerCase() === exercise_name.toLowerCase());
-        if (idx === -1) return err(`Exercise "${exercise_name}" not found in ${phase}.`);
-        if (!updates || Object.keys(updates).length === 0) {
-          return err("No fields provided to update. Specify at least one of: exercise, sets, reps, weight_lbs, notes.");
-        }
-        const expectedExercise = { ...arr[idx], ...updates };
-        arr[idx] = expectedExercise;
-
-        const { data: updated, error: upErr } = await supabase
-          .from("workout_plans")
-          .update({ [phase]: arr, updated_at: new Date().toISOString() })
-          .eq("user_id", userId)
-          .eq("date", date)
-          .select()
-          .single();
-        if (upErr) return err(upErr.message);
-        if (!updated) return err("Workout plan update returned no row — change may not have been saved.");
-
-        // Read-after-write verification (#319 / #239 pattern): re-read the row
-        // and confirm the matched exercise actually contains the requested
-        // updates. Catches the silent-mismatch class where the update appears
-        // to succeed but the row reads back unchanged.
-        const verifiedArr = (updated[phase] as WorkoutExercise[]) ?? [];
-        const verifiedExercise = verifiedArr[idx];
-        const fieldChecks = Object.entries(updates).filter(
-          ([key, value]) => (verifiedExercise as unknown as Record<string, unknown>)[key] !== value
-        );
-        if (fieldChecks.length > 0) {
-          return err(
-            `Workout plan update returned but the row read-back doesn't show the change: ${fieldChecks
-              .map(([k, v]) => `${k}=${JSON.stringify(v)}`)
-              .join(", ")}`
-          );
-        }
-
-        let calendar_synced: boolean | null = null;
-        let calendar_error: string | undefined;
-        if (updated.calendar_event_id) {
-          try {
-            const auth = getGoogleAuthClient();
-            const calendar = google.calendar({ version: "v3", auth });
-            const description = buildCalendarDescription(updated.warmup, updated.workout, updated.cooldown);
-            await withTimeout(
-              calendar.events.patch({
-                calendarId: "primary",
-                eventId: updated.calendar_event_id,
-                requestBody: { description },
-              }),
-              8000,
-              "calendar patch"
-            );
-            calendar_synced = true;
-          } catch (calErr) {
-            calendar_synced = false;
-            calendar_error = calErr instanceof Error ? calErr.message : "Calendar description sync failed";
-          }
-        }
-
-        return ok({ plan: updated, calendar_synced, ...(calendar_error ? { calendar_error } : {}) });
-      },
-    }),
-
-    get_workout_history: tool({
-      description: "Fetch logged strength-session performance (actual sets/reps/weights/RPE/notes) for progression analysis. Returns recent sessions scoped to the user, with per-set detail. Use this before suggesting progression for any exercise. Weights are returned in kg canonically — call get_profile for the user's weight_unit preference before reporting numbers.",
-      inputSchema: jsonSchema<{ exercise_name?: string; days?: number }>({
-        type: "object",
-        properties: {
-          exercise_name: {
-            type: "string",
-            description: "Optional exercise name (case-insensitive partial match). Omit to get the full recent history across all exercises.",
-          },
-          days: {
-            type: "number",
-            description: "Number of days back to include. Defaults to 30.",
-          },
-        },
-      }),
-      execute: async ({ exercise_name, days = 30 }) => {
-        if (!userId) return { error: "Not authenticated" };
-        if (isDemo) return { demo: true, note: "Demo mode — no real strength history.", sessions: [] };
-
-        const sinceStr = daysAgoString(Math.max(1, Math.round(days)));
-        const { data: sessions, error } = await supabase
-          .from("strength_sessions")
-          .select("id, performed_on, started_at, completed_at, perceived_effort, notes, sets:strength_session_sets(exercise_name, exercise_order, set_number, weight_kg, reps, rpe, notes)")
-          .eq("user_id", userId)
-          .gte("performed_on", sinceStr)
-          .order("performed_on", { ascending: false })
-          .limit(30);
-        if (error) return { error: error.message };
-
-        interface SetRow {
-          exercise_name: string;
-          exercise_order: number;
-          set_number: number;
-          weight_kg: number | null;
-          reps: number | null;
-          rpe: number | null;
-          notes: string | null;
-        }
-        interface SessionRow {
-          id: string;
-          performed_on: string;
-          started_at: string | null;
-          completed_at: string | null;
-          perceived_effort: number | null;
-          notes: string | null;
-          sets: SetRow[];
-        }
-
-        const rows = (sessions ?? []) as SessionRow[];
-        const filtered = exercise_name
-          ? rows
-              .map((s) => ({
-                ...s,
-                sets: s.sets.filter((set) =>
-                  set.exercise_name.toLowerCase().includes(exercise_name.toLowerCase())
-                ),
-              }))
-              .filter((s) => s.sets.length > 0)
-          : rows;
-
-        return {
-          window_days: days,
-          exercise_filter: exercise_name ?? null,
-          weight_unit: "kg",
-          session_count: filtered.length,
-          sessions: filtered.map((s) => ({
-            session_id: s.id,
-            performed_on: s.performed_on,
-            started_at: s.started_at,
-            completed_at: s.completed_at,
-            perceived_effort: s.perceived_effort,
-            notes: s.notes,
-            sets: s.sets
-              .slice()
-              .sort((a, b) =>
-                a.exercise_order !== b.exercise_order
-                  ? a.exercise_order - b.exercise_order
-                  : a.set_number - b.set_number
-              ),
-          })),
-        };
-      },
-    }),
-
-    get_stock_quote: tool({
-      description: "Get current price and daily change for a stock ticker. Checks stocks_cache first; falls back to a live Polygon.io fetch if the cache is stale (>6h) or the ticker is not cached. Use when the user asks about a stock price or market move.",
-      inputSchema: jsonSchema<{ ticker: string }>({
-        type: "object",
-        required: ["ticker"],
-        properties: {
-          ticker: { type: "string", description: "Stock ticker symbol, e.g. AAPL, NVDA, BTC-USD" },
-        },
-      }),
-      execute: async ({ ticker }) => {
-        const sym = ticker.toUpperCase();
-
-        // 1. Check stocks_cache
-        const { data: cached } = await supabase
-          .from("stocks_cache")
-          .select("ticker,price,change_abs,change_pct,fetched_at")
-          .eq("user_id", userId)
-          .eq("ticker", sym)
-          .maybeSingle();
-
-        const sixHoursAgo = new Date(Date.now() - 6 * 60 * 60 * 1000).toISOString();
-        if (cached && cached.fetched_at > sixHoursAgo) {
-          return {
-            ticker: cached.ticker,
-            price: cached.price,
-            change_abs: cached.change_abs,
-            change_pct: cached.change_pct,
-            fetched_at: cached.fetched_at,
-            source: "cache",
-          };
-        }
-
-        // 2. Live fetch from Polygon
-        const apiKey = process.env.POLYGON_API_KEY;
-        if (!apiKey) return { error: "POLYGON_API_KEY not configured" };
-
-        const res = await fetch(
-          `https://api.polygon.io/v2/aggs/ticker/${sym}/prev?apiKey=${apiKey}`,
-          { cache: "no-store" },
-        );
-        if (!res.ok) return { error: `Polygon returned ${res.status}` };
-
-        const json = await res.json() as { results?: { o: number; c: number }[] };
-        const bar = json.results?.[0];
-        if (!bar) return { error: "No data returned for this ticker" };
-
-        const price = bar.c;
-        const changeAbs = bar.c - bar.o;
-        const changePct = bar.o !== 0 ? ((bar.c - bar.o) / bar.o) * 100 : 0;
-
-        return {
-          ticker: sym,
-          price,
-          change_abs: changeAbs,
-          change_pct: changePct,
-          fetched_at: new Date().toISOString(),
-          source: "live",
-        };
-      },
-    }),
-
-    get_sports_data: tool({
-      description: "Get next game, last result, or standings for one of the user's favorite sports teams. Reads sports_cache first; falls back to a live TheSportsDB fetch when the cache is older than 12h or when the requested game is within 24h of now. Use when the user asks about a team's schedule, recent result, or standings.",
-      inputSchema: jsonSchema<{
-        team_name?: string;
-        team_id?: string;
-        query_type: "next_game" | "last_result" | "standings";
-      }>({
-        type: "object",
-        required: ["query_type"],
-        properties: {
-          team_name: { type: "string", description: "Team name as the user said it (e.g. 'Warriors'). Matched case-insensitively against the user's favorites." },
-          team_id: { type: "string", description: "TheSportsDB team_id, if known (preferred over team_name)." },
-          query_type: { type: "string", enum: ["next_game", "last_result", "standings"] },
-        },
-      }),
-      execute: async ({ team_name, team_id, query_type }) => {
-        const { data: favRow } = await supabase
-          .from("profile")
-          .select("value")
-          .eq("user_id", userId)
-          .eq("key", "sports_favorites")
-          .maybeSingle();
-
-        type Fav = { team_id: string; name: string; league: string; league_id: string; badge: string | null; color: string | null };
-        const favorites: Fav[] = favRow?.value ? (JSON.parse(favRow.value) as Fav[]) : [];
-        if (favorites.length === 0) {
-          return { error: "No favorite teams configured. Add some in Settings." };
-        }
-
-        const fav = team_id
-          ? favorites.find((f) => f.team_id === team_id)
-          : favorites.find((f) => f.name.toLowerCase().includes((team_name ?? "").toLowerCase()));
-        if (!fav) {
-          return { error: `Team not in favorites. Available: ${favorites.map((f) => f.name).join(", ")}` };
-        }
-
-        const { data: cached } = await supabase
-          .from("sports_cache")
-          .select("data,fetched_at")
-          .eq("user_id", userId)
-          .eq("team_id", fav.team_id)
-          .maybeSingle();
-
-        const now = Date.now();
-        const cacheAgeHrs = cached ? (now - new Date(cached.fetched_at).getTime()) / 3_600_000 : Infinity;
-
-        // Decide whether to live-fetch
-        let needsLive = !cached || cacheAgeHrs > 12;
-        if (cached && !needsLive) {
-          const data = cached.data as SportsCacheData;
-          if (query_type === "next_game") {
-            const next = data.upcoming?.[0];
-            if (next && Math.abs(new Date(next.start_time).getTime() - now) < 24 * 3_600_000) needsLive = true;
-          }
-          if (query_type === "last_result") {
-            const last = data.recent?.[0];
-            if (last && now - new Date(last.start_time).getTime() < 24 * 3_600_000) needsLive = true;
-          }
-        }
-
-        let data: SportsCacheData;
-        let source: "cache" | "live" | "stale-cache" = "cache";
-
-        if (needsLive) {
-          try {
-            await syncSports(supabase, userId, [fav]);
-            const { data: fresh } = await supabase
-              .from("sports_cache")
-              .select("data")
-              .eq("user_id", userId)
-              .eq("team_id", fav.team_id)
-              .maybeSingle();
-            data = (fresh?.data ?? cached?.data) as SportsCacheData;
-            source = "live";
-          } catch (e) {
-            if (!cached) return { error: `Live fetch failed and no cache: ${(e as Error).message}` };
-            data = cached.data as SportsCacheData;
-            source = "stale-cache";
-          }
-        } else {
-          data = cached!.data as SportsCacheData;
-        }
-
-        if (query_type === "next_game") {
-          return { team: fav.name, league: fav.league, next_game: data.upcoming?.[0] ?? null, source };
-        }
-        if (query_type === "last_result") {
-          return { team: fav.name, league: fav.league, last_result: data.recent?.[0] ?? null, source };
-        }
-        return { team: fav.name, league: fav.league, standings: data.standings ?? null, source };
-      },
-    }),
+    ...buildTasksTools(toolContext),
+    ...buildHabitsTools(toolContext),
+    ...buildFitnessTools(toolContext),
+    ...buildProfileTools(toolContext),
+    ...buildGmailTools(toolContext),
+    ...buildCalendarTools(toolContext),
+    ...buildMealsTools(toolContext),
+    ...buildSessionTools(toolContext),
+    ...buildWorkoutTools(toolContext),
+    ...buildStocksTools(toolContext),
+    ...buildSportsTools(toolContext),
   };
 
   // Select model: demo → Groq Llama (free tier); real user → Anthropic tier selection
