+++ web/src/lib/tools/habits.ts
@@ -1,0 +1,75 @@
+import { tool, jsonSchema } from "ai";
+import { todayString } from "@/lib/timezone";
+import { ok, err } from "./_contract";
+import type { ToolContext } from "./_context";
+
+export function buildHabitsTools({ supabase, userId }: ToolContext) {
+  return {
+    get_habits_today: tool({
+      description: "Get all active habits and their completion status for today (or a specified date).",
+      inputSchema: jsonSchema<{ date?: string }>({
+        type: "object",
+        properties: {
+          date: { type: "string", description: "Date in YYYY-MM-DD format. Defaults to today." },
+        },
+      }),
+      execute: async ({ date }) => {
+        const targetDate = date ?? todayString();
+        let regQ = supabase.from("habit_registry").select("id, name, emoji, category").eq("active", true);
+        let logQ = supabase.from("habits").select("habit_id, completed, notes").eq("date", targetDate);
+        if (userId) { regQ = regQ.eq("user_id", userId); logQ = logQ.eq("user_id", userId); }
+        const [registryResult, logsResult] = await Promise.all([regQ, logQ]);
+        if (registryResult.error) return { error: registryResult.error.message };
+        const logMap = new Map(
+          (logsResult.data ?? []).map((l: { habit_id: string; completed: boolean; notes: string | null }) => [l.habit_id, l])
+        );
+        return (registryResult.data ?? []).map((h: { id: string; name: string; emoji: string | null; category: string | null }) => ({
+          id: h.id,
+          name: h.name,
+          emoji: h.emoji,
+          category: h.category,
+          completed: logMap.get(h.id)?.completed ?? false,
+          notes: logMap.get(h.id)?.notes ?? null,
+        }));
+      },
+    }),
+
+    log_habit: tool({
+      description: "Log a habit as completed. Looks up the habit by name from the registry.",
+      inputSchema: jsonSchema<{ name: string; date?: string; notes?: string }>({
+        type: "object",
+        required: ["name"],
+        properties: {
+          name: { type: "string", description: "Habit name (case-insensitive partial match)." },
+          date: { type: "string", description: "Date in YYYY-MM-DD format. Defaults to today." },
+          notes: { type: "string", description: "Optional notes." },
+        },
+      }),
+      execute: async ({ name, date, notes }) => {
+        const targetDate = date ?? todayString();
+        let habQ = supabase
+          .from("habit_registry")
+          .select("id, name")
+          .ilike("name", `%${name}%`)
+          .eq("active", true)
+          .limit(1);
+        if (userId) habQ = habQ.eq("user_id", userId);
+        const { data: habits, error: lookupError } = await habQ;
+        if (lookupError) return err(lookupError.message);
+        if (!habits || habits.length === 0) return err(`No active habit matching "${name}" found.`);
+        const habit = habits[0];
+        const { data, error: upsertError } = await supabase
+          .from("habits")
+          .upsert(
+            { user_id: userId, habit_id: habit.id, date: targetDate, completed: true, notes: notes ?? null },
+            { onConflict: "habit_id,date" }
+          )
+          .select("habit_id, date, completed, notes")
+          .single();
+        if (upsertError) return err(upsertError.message);
+        if (!data) return err("Upsert returned no row — habit may not have been logged.");
+        return ok({ habit: habit.name, log: data });
+      },
+    }),
+  };
+}
